Move top-level types back into registry mod, and reexport
These types were moved to the top level because it was becoming hard to pick them out from the considerably sized registry module. Now that the parsing has been moved into a submodule module, it is probably clearer to have the registry types grouped together.
